refactor: Unchecked Exception 추가
