+++ app/src/main/java/com/techcourse/dao/UserDao.java
@@ -1,6 +1,9 @@
 package com.techcourse.dao;
 
 import com.techcourse.domain.User;
+import nextstep.jdbc.JdbcTemplate;
+import nextstep.jdbc.PreparedStatementSetter;
+import nextstep.jdbc.RowMapper;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -15,22 +18,15 @@ public class UserDao {
 
     private static final Logger log = LoggerFactory.getLogger(UserDao.class);
 
-    private final DataSource dataSource;
+    private final JdbcTemplate jdbcTemplate;
 
     public UserDao(DataSource dataSource) {
-        this.dataSource = dataSource;
+        this.jdbcTemplate = new JdbcTemplate(dataSource);
     }
 
     public void insert(User user) {
         final String sql = "insert into users (account, password, email) values (?, ?, ?)";
-        JdbcTemplate insertJdbcTemplate = new JdbcTemplate() {
-            @Override
-            DataSource getDataSource() {
-                return dataSource;
-            }
-        };
-
-        insertJdbcTemplate.update(sql, new PreparedStatementSetter() {
+        jdbcTemplate.update(sql, new PreparedStatementSetter() {
             @Override
             public void setValues(PreparedStatement pstmt) throws SQLException {
                 pstmt.setString(1, user.getAccount());
@@ -42,14 +38,8 @@ public class UserDao {
 
     public void update(User user) {
         final String sql = "update users set account=?, password=?, email=? where id=?";
-        JdbcTemplate updateJdbcTemplate = new JdbcTemplate() {
-            @Override
-            DataSource getDataSource() {
-                return dataSource;
-            }
-        };
 
-        updateJdbcTemplate.update(sql, new PreparedStatementSetter() {
+        jdbcTemplate.update(sql, new PreparedStatementSetter() {
             @Override
             public void setValues(PreparedStatement pstmt) throws SQLException {
                 pstmt.setString(1, user.getAccount());
@@ -62,14 +52,8 @@ public class UserDao {
 
     public List<User> findAll() {
         final String sql = "select id, account, password, email from users";
-        JdbcTemplate selectAllJdbcTemplate = new JdbcTemplate() {
-            @Override
-            DataSource getDataSource() {
-                return dataSource;
-            }
-        };
 
-        return (List<User>) selectAllJdbcTemplate.query(sql, new PreparedStatementSetter() {
+        return (List<User>) jdbcTemplate.query(sql, new PreparedStatementSetter() {
             @Override
             public void setValues(PreparedStatement pstmt) throws SQLException {
             }
@@ -92,14 +76,8 @@ public class UserDao {
 
     public User findById(Long id) {
         final String sql = "select id, account, password, email from users where id = ?";
-        JdbcTemplate selectByIdJdbcTemplate = new JdbcTemplate() {
-            @Override
-            DataSource getDataSource() {
-                return dataSource;
-            }
-        };
 
-        return (User) selectByIdJdbcTemplate.query(sql, new PreparedStatementSetter() {
+        return (User) jdbcTemplate.query(sql, new PreparedStatementSetter() {
             @Override
             public void setValues(PreparedStatement pstmt) throws SQLException {
                 pstmt.setLong(1, id);
@@ -121,14 +99,8 @@ public class UserDao {
 
     public User findByAccount(String account) {
         final String sql = "select id, account, password, email from users where account = ?";
-        JdbcTemplate selectByIdJdbcTemplate = new JdbcTemplate() {
-            @Override
-            DataSource getDataSource() {
-                return dataSource;
-            }
-        };
 
-        return (User) selectByIdJdbcTemplate.query(sql, new PreparedStatementSetter() {
+        return (User) jdbcTemplate.query(sql, new PreparedStatementSetter() {
             @Override
             public void setValues(PreparedStatement pstmt) throws SQLException {
                 pstmt.setString(1, account);
