+++ jdbc/src/main/java/nextstep/exception/DataAccessException.java
@@ -1,0 +1,11 @@
+package nextstep.exception;
+
+public class DataAccessException  extends RuntimeException {
+    public DataAccessException(String message) {
+        super(message);
+    }
+
+    public DataAccessException(Throwable cause) {
+        super(cause);
+    }
+}
