+++ jdbc/src/main/java/nextstep/jdbc/RowMapper.java
@@ -1,4 +1,4 @@
-package com.techcourse.dao;
+package nextstep.jdbc;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
